requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
opencv-python-headless>=4.5.3
numpy>=1.21.2
pillow>=8.3.2
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tensorflow-macos>=2.13.0
tensorflow-metal>=0.7.0
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
imagehash>=4.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
requests>=2.26.0
scikit-learn>=1.0.2
scikit-image>=0.18.3
tqdm>=4.62.3 